# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EcoFirma API Gateway
  version: 1.0.0
  description: REST API exposed by the EcoFirma API Gateway for users, documents and asynchronous signatures.
servers:
  - url: http://localhost:8081
    description: Local Docker deployment
...
paths:
  /api/users/register:
    post:
      tags: [Users Service]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/users/login:
    post:
      tags: [Users Service]
      summary: Authenticate a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: JWT token and user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/documents:
    get:
      tags: [Documents Service]
      summary: List documents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'
    post:
      tags: [Documents Service]
      summary: Create a document
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentCreateRequest'
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    UserRegisterRequest:
      type: object
      required: [nombre, email, password]
      properties:
        nombre:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    UserLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    UserResponse:
      type: object
      properties:
        userId:
          type: string
        nombre:
          type: string
        email:
          type: string
          format: email
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserResponse'
    DocumentCreateRequest:
      type: object
      required: [titulo, contenidoBase64, autorId]
      properties:
        titulo:
          type: string
        contenidoBase64:
          type: string
        autorId:
          type: string
...
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
        titulo:
          type: string
        contenidoBase64:
          type: string
        autorId:
          type: string
        estado:
          $ref: '#/components/schemas/DocumentStatus'
        status:
          $ref: '#/components/schemas/DocumentStatus'
    DocumentStatus:
      type: string
      enum: [PENDIENTE, EN_PROCESO, FIRMADO, ERROR]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string